#55 refactor: move main app-files
