+++ src/__tests__/Control.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import ControlApp from '../Control';
+import ControlApp from './../ControlComponents/ControlApp';
 
 test('renders correctly', () => {
   const tree = renderer.create(<ControlApp />).toJSON();
